Refactor native builtin registration
