+++ include/builtins.h
@@ -1,0 +1,8 @@
+#ifndef ORUS_BUILTINS_H
+#define ORUS_BUILTINS_H
+
+#include "vm.h"
+
+void initBuiltins(void);
+
+#endif // ORUS_BUILTINS_H
